+++ weblate/checks/tests/test_same_checks.py
@@ -42,6 +42,7 @@ class SameCheckTest(CheckTestCase):
         unit = MockUnit(code='en')
         # Is template
         unit.translation.is_template = True
+        unit.translation.is_source = True
         self.assertTrue(self.check.should_skip(unit))
         # Is same as source
         unit.translation.template = False