+++ weblate/trans/views/js.py
@@ -34,7 +34,7 @@ from weblate.machinery import MACHINE_TRANSLATION_SERVICES
 from weblate.machinery.base import MachineTranslationError
 from weblate.screenshots.forms import ScreenshotForm
 from weblate.trans.forms import CheckFlagsForm, ContextForm
-from weblate.trans.models import Change, Source, Unit
+from weblate.trans.models import Change, Unit
 from weblate.trans.util import sort_objects
 from weblate.utils.celery import get_task_progress, is_task_ready
 from weblate.utils.errors import report_error
@@ -161,20 +161,20 @@ def ignore_check(request, check_id):
 def ignore_check_source(request, check_id, pk):
     obj = get_object_or_404(Check, pk=int(check_id))
     request.user.check_access(obj.project)
-    source = get_object_or_404(Source, pk=int(pk))
+    source = get_object_or_404(Unit, pk=int(pk))
 
-    if (obj.project != source.component.project
+    if (obj.project != source.translation.component.project
             or not request.user.has_perm('unit.check', obj.project)
-            or not request.user.has_perm('source.edit', source.component)):
+            or not request.user.has_perm('source.edit', source.translation.component)):
         raise PermissionDenied()
 
     # Mark check for ignoring
     ignore = obj.check_obj.ignore_string
-    if ignore not in source.check_flags:
-        if source.check_flags:
-            source.check_flags += ', {}'.format(ignore)
+    if ignore not in source.extra_flags:
+        if source.extra_flags:
+            source.extra_flags += ', {}'.format(ignore)
         else:
-            source.check_flags = ignore
+            source.extra_flags = ignore
         source.save()
 
     # response for AJAX
@@ -299,11 +299,11 @@ def get_detail(request, project, component, checksum):
             'project': component.project,
             'next': request.GET.get('next', ''),
             'context_form': ContextForm(
-                initial={'context': source.context}
+                initial={'context': source.extra_context}
             ),
 
             'check_flags_form': CheckFlagsForm(
-                initial={'flags': source.check_flags}
+                initial={'flags': source.extra_flags}
             ),
             'screenshot_form': ScreenshotForm(),
             'extra_flags': PLAIN_FLAGS.items(),